Update to use Zustand (#1)
* Replaced useQuery with Zustand

+++ src/api/supabase/images/deleteImage.ts
@@ -1,8 +1,23 @@
 import { client } from "..";
-import { Image } from "../../types";
+import { ImageWithProjects } from "../../../types";
 
-const deleteImage = async (image: Image): Promise<Image> => {
+const deleteImage = async (
+  image: ImageWithProjects,
+): Promise<ImageWithProjects> => {
   if (image) {
+    const { error: linksError, status: linksStatus } = await client
+      .from("projects_images")
+      .delete()
+      .eq("image_id", image.id);
+
+    if (linksError) {
+      throw linksError;
+    }
+
+    if (linksStatus !== 204) {
+      throw new Error("Could not delete image");
+    }
+
     const { error, status } = await client
       .from("images")
       .delete()
