+++ src/api/supabase/projects/deleteProject.ts
@@ -1,8 +1,23 @@
-import { Project } from "../../types";
+import { ProjectWithImages } from "../../../types";
 import { client } from "../client";
 
-const deleteProject = async (project: Project): Promise<Project> => {
+const deleteProject = async (
+  project: ProjectWithImages,
+): Promise<ProjectWithImages> => {
   if (project) {
+    const { error: linksError, status: linksStatus } = await client
+      .from("projects_images")
+      .delete()
+      .eq("project_id", project.id);
+
+    if (linksError) {
+      throw linksError;
+    }
+
+    if (linksStatus !== 204) {
+      throw new Error("Could not delete project");
+    }
+
     const { error, status } = await client
       .from("projects")
       .delete()
